Day 5 Part 1 - TDD REFACTOR
- add extra test
- change from int to long to accomodate the larger input

+++ Year 2023/src/main/java/com/github/silviuburceadev/aoc/garden/DestSrcRange.java
@@ -2,7 +2,7 @@ package com.github.silviuburceadev.aoc.garden;
 
 import static java.lang.Integer.parseInt;
 
-public record DestSrcRange(int destination, int source, int range) {
+public record DestSrcRange(long destination, long source, long range) {
     public DestSrcRange(String input) {
         this(input.split("\\s+"));
     }
@@ -11,7 +11,7 @@ public record DestSrcRange(int destination, int source, int range) {
         this(parseInt(input[0]), parseInt(input[1]), parseInt(input[2]));
     }
 
-    public int apply(int seed) {
+    public long apply(long seed) {
         if (seed < source || seed >= source + range) return seed;
         return seed + (destination - source);
     }
